+++ src/main/java/me/shiiyuko/manosaba/init/ManosabaSounds.java
@@ -28,6 +28,22 @@ public final class ManosabaSounds {
             SOUND_EVENTS.register("button_click_start_game", () -> SoundEvent.createVariableRangeEvent(
                     ResourceLocation.fromNamespaceAndPath(Manosaba.MODID, "button_click_start_game")));
 
+    public static final DeferredHolder<SoundEvent, SoundEvent> SFX_SYSTEM_SUBMIT =
+            SOUND_EVENTS.register("sfx_system_submit_001", () -> SoundEvent.createVariableRangeEvent(
+                    ResourceLocation.fromNamespaceAndPath(Manosaba.MODID, "sfx_system_submit_001")));
+
+    public static final DeferredHolder<SoundEvent, SoundEvent> SFX_SYSTEM_CANCEL =
+            SOUND_EVENTS.register("sfx_system_cancel_001", () -> SoundEvent.createVariableRangeEvent(
+                    ResourceLocation.fromNamespaceAndPath(Manosaba.MODID, "sfx_system_cancel_001")));
+
+    public static final DeferredHolder<SoundEvent, SoundEvent> SFX_SYSTEM_LOADDATA =
+            SOUND_EVENTS.register("sfx_system_loaddata_001", () -> SoundEvent.createVariableRangeEvent(
+                    ResourceLocation.fromNamespaceAndPath(Manosaba.MODID, "sfx_system_loaddata_001")));
+
+    public static final DeferredHolder<SoundEvent, SoundEvent> SFX_SYSTEM_STARTGAME =
+            SOUND_EVENTS.register("sfx_system_startgame_001", () -> SoundEvent.createVariableRangeEvent(
+                    ResourceLocation.fromNamespaceAndPath(Manosaba.MODID, "sfx_system_startgame_001")));
+
     private ManosabaSounds() {
     }
 }